Everything compiles
