+++ rudder-core/src/main/scala/com/normation/rudder/services/modification/ModificationService.scala
@@ -37,16 +37,15 @@
 
 package com.normation.rudder.services.modification
 
-import com.normation.rudder.repository._
 import com.normation.eventlog._
-import net.liftweb.common._
-import com.normation.utils.StringUuidGenerator
-import org.eclipse.jgit.lib.PersonIdent
+
+import com.normation.rudder.repository._
 import com.normation.rudder.repository.EventLogRepository
-import scala.concurrent.Await
-import scala.util.{Try, Success => TSuccess, Failure => TFailure}
-import scala.concurrent.duration.Duration
-import java.util.concurrent.TimeUnit
+import com.normation.utils.StringUuidGenerator
+
+import org.eclipse.jgit.lib.PersonIdent
+
+import net.liftweb.common._
 
 class ModificationService(
       eventLogRepository : EventLogRepository
@@ -54,22 +53,20 @@ class ModificationService(
     , itemArchiveManager : ItemArchiveManager
     , uuidGen : StringUuidGenerator ) {
 
-  def getCommitsfromEventLog(eventLog:EventLog) : Option[GitCommitId] = {
+  def getCommitsfromEventLog(eventLog:EventLog) : Box[Option[GitCommitId]] = {
     eventLog.modificationId match {
-      case None        => None
-      case Some(modId) =>
-        Try(Await.result(gitModificationRepository.getCommits(modId), Duration.Inf)) match {
-        case TSuccess(Full(s))     => s.headOption
-        case TSuccess(eb:EmptyBox) => None
-        case TFailure(ex) => Failure(s"Error when finding back commit ID for modification ID '${modId.value}': ${ex.getMessage}", Full(ex), Empty)
-      }
+      case None        => Full(None)
+      case Some(modId) => gitModificationRepository.getCommits(modId)
     }
   }
 
   def restoreToEventLog(eventLog:EventLog, commiter:PersonIdent, rollbackedEvents:Seq[EventLog], target:EventLog) = {
     for {
-      commit   <-  getCommitsfromEventLog(eventLog)
-      rollback <- itemArchiveManager.rollback(commit, commiter, ModificationId(uuidGen.newUuid), eventLog.principal, None, rollbackedEvents, target, "after", false)
+      commit   <- getCommitsfromEventLog(eventLog)
+      rollback <- commit match {
+                    case None    => Failure(s"The event log ${eventLog.id.getOrElse("")} don't have a matching commit ID and can't be restored")
+                    case Some(x) => itemArchiveManager.rollback(x, commiter, ModificationId(uuidGen.newUuid), eventLog.principal, None, rollbackedEvents, target, "after", false)
+                  }
     } yield {
       rollback
     }
@@ -78,11 +75,15 @@ class ModificationService(
 
   def restoreBeforeEventLog(eventLog:EventLog, commiter:PersonIdent, rollbackedEvents:Seq[EventLog], target:EventLog) = {
     for {
-      commit   <-  getCommitsfromEventLog(eventLog)
-      parentCommit = GitCommitId(commit.value+"^")
-      rollback <- itemArchiveManager.rollback(parentCommit, commiter, ModificationId(uuidGen.newUuid), eventLog.principal, None, rollbackedEvents, target, "before", false)
-    } yield { val parentCommit = GitCommitId(commit.value+"^")
-        rollback
+      commit   <- getCommitsfromEventLog(eventLog)
+      rollback <- commit match {
+                    case None    => Failure(s"The event log ${eventLog.id.getOrElse("")} don't have a matching commit ID and can't be restored")
+                    case Some(x) =>
+                      val parentCommit = GitCommitId(x.value+"^")
+                      itemArchiveManager.rollback(parentCommit, commiter, ModificationId(uuidGen.newUuid), eventLog.principal, None, rollbackedEvents, target, "before", false)
+                  }
+    } yield {
+      rollback
     }
   }
 }
