+++ rudder-core/src/test/scala/com/normation/rudder/repository/jdbc/HistorizationRepositoryTest.scala
@@ -59,7 +59,7 @@ import org.specs2.runner.JUnitRunner
 @RunWith(classOf[JUnitRunner])
 class HistorizationRepositoryTest extends DBCommon with BoxSpecMatcher  {
 
-  val repos = new HistorizationJdbcRepository(schema)
+  val repos = new HistorizationJdbcRepository(doobie)
   val service = new HistorizationServiceImpl(repos)
