+++ rudder-web/src/main/scala/bootstrap/liftweb/AppConfig.scala
@@ -127,7 +127,6 @@ import com.normation.rudder.migration._
 import com.normation.rudder.web.rest.parameter._
 import com.normation.rudder.web.rest.changeRequest._
 import com.normation.rudder.reports.execution._
-import com.normation.rudder.reports.statusUpdate._
 import com.normation.rudder.domain.policies.RuleId
 import com.normation.rudder.appconfig._
 import com.normation.rudder.rule.category._
@@ -147,7 +146,7 @@ import org.apache.commons.io.FileUtils
 import com.normation.templates.FillTemplatesService
 
 import com.normation.rudder.web.rest.technique._
-import com.normation.rudder.db.SlickSchema
+import com.normation.rudder.db.Doobie
 
 /**
  * Define a resource for configuration.
@@ -1051,7 +1050,7 @@ object RudderConfig extends Loggable {
       , "default-directive-names.conf"
     )
   }
-  private[this] lazy val historizationJdbcRepository = new HistorizationJdbcRepository(slickSchema)
+  private[this] lazy val historizationJdbcRepository = new HistorizationJdbcRepository(doobie)
   private[this] lazy val startStopOrchestratorImpl: StartStopOrchestrator = {
     if (!(new File(BIN_EMERGENCY_STOP)).exists)
       ApplicationLogger.error("The 'red button' program is not present at: '%s'. You will experience error when trying to use that functionnality".format(BIN_EMERGENCY_STOP))
@@ -1144,7 +1143,7 @@ object RudderConfig extends Loggable {
   private[this] lazy val woParameterServiceImpl = new WoParameterServiceImpl(roParameterServiceImpl, woLDAPParameterRepository, asyncDeploymentAgentImpl)
 
   ///// items archivers - services that allows to transform items to XML and save then on a Git FS /////
-  private[this] lazy val gitModificationRepository = new GitModificationRepositoryImpl(slickSchema)
+  private[this] lazy val gitModificationRepository = new GitModificationRepositoryImpl(doobie)
   private[this] lazy val gitRuleArchiver: GitRuleArchiver = new GitRuleArchiverImpl(
       gitRepo
     , new File(RUDDER_DIR_GITROOT)
@@ -1501,7 +1500,7 @@ object RudderConfig extends Loggable {
   private[this] lazy val updateExpectedRepo = new UpdateExpectedReportsJdbcRepository(jdbcTemplate, transactionManager, findExpectedRepo, findExpectedRepo)
   private[this] lazy val reportsRepositoryImpl = new ReportsJdbcRepository(jdbcTemplate)
   private[this] lazy val dataSourceProvider = new RudderDatasourceProvider(RUDDER_JDBC_DRIVER, RUDDER_JDBC_URL, RUDDER_JDBC_USERNAME, RUDDER_JDBC_PASSWORD, RUDDER_JDBC_MAX_POOL_SIZE)
-  private[this] lazy val slickSchema = new SlickSchema(dataSourceProvider.datasource)
+  private[this] lazy val doobie = new Doobie(dataSourceProvider.datasource)
   private[this] lazy val jdbcTemplate = {
     val template = new org.springframework.jdbc.core.JdbcTemplate(dataSourceProvider.datasource)
     template
@@ -1652,7 +1651,7 @@ object RudderConfig extends Loggable {
    * Event log migration
    */
 
-  private[this] lazy val migrationRepository = new MigrationEventLogRepository(slickSchema)
+  private[this] lazy val migrationRepository = new MigrationEventLogRepository(doobie)
 
   private[this] lazy val eventLogsMigration_2_3 = new EventLogsMigration_2_3(
       jdbcTemplate        = jdbcTemplate
@@ -1820,9 +1819,7 @@ object RudderConfig extends Loggable {
     , roLdapDirectiveRepository
     , reportingServiceImpl
     , techniqueRepositoryImpl)
-  private[this] lazy val propertyRepository = new RudderPropertiesRepositoryImpl(
-      slickSchema
-    , reportsRepository )
+  private[this] lazy val propertyRepository = new RudderPropertiesRepositoryImpl(doobie)
   private[this] lazy val autoReportLogger = new AutomaticReportLogger(
       propertyRepository
     , reportsRepositoryImpl
@@ -1839,14 +1836,14 @@ object RudderConfig extends Loggable {
    * Agent runs: we use a cache for them.
    */
   private[this] lazy val cachedAgentRunRepository = {
-    val roRepo = new RoReportsExecutionRepositoryImpl(slickSchema, pgIn)
+    val roRepo = new RoReportsExecutionRepositoryImpl(doobie, pgIn)
     new CachedReportsExecutionRepository(
         roRepo
-      , new WoReportsExecutionRepositoryImpl(slickSchema, roRepo )
+      , new WoReportsExecutionRepositoryImpl(doobie, roRepo )
     )
   }
 
-  val updatesEntryJdbcRepository = new SlickStatusUpdateRepository(slickSchema)
+  val updatesEntryJdbcRepository = new LastProcessedReportRepositoryImpl(doobie)
 
   val executionService = {
     val max   = if (RUDDER_REPORTS_EXECUTION_MAX_DAYS > 0) {
